screens separated in dir
