+++ js/main.js
@@ -1,4 +1,4 @@
-import {makeScreenActive} from './util';
-import insertIntro from './intro';
+import {makeScreenActive} from './util.js';
+import insertIntro from './screens/intro.js';
 
 makeScreenActive(insertIntro());
