+++ js/screens/guessForEach.js
@@ -1,8 +1,8 @@
-import {getFromTemplate, makeScreenActive} from './util.js';
-import insertFindImgOrPhoto from './findImgOrPhoto';
+import {getFromTemplate, makeScreenActive} from '../util.js';
+import insertGuessForOne from './guessForOne.js';
 import insertGreeting from './greeting.js';
 
-export default function insertGuessForOne() {
+export default function insertGuessForEach() {
   const node = getFromTemplate(`
     <header class="header">
       <button class="back">
@@ -16,22 +16,33 @@ export default function insertGuessForOne() {
       </button>
       <div class="game__timer">NN</div>
       <div class="game__lives">
-        <img src="img/heart__empty.svg" class="game__heart" alt="Life" width="31" height="27">
+        <img src="img/heart__empty.svg" class="game__heart" alt=" Missed Life" width="31" height="27">
         <img src="img/heart__full.svg" class="game__heart" alt="Life" width="31" height="27">
         <img src="img/heart__full.svg" class="game__heart" alt="Life" width="31" height="27">
       </div>
     </header>
     <section class="game">
-      <p class="game__task">Угадай, фото или рисунок?</p>
-      <form class="game__content  game__content--wide">
+      <p class="game__task">Угадайте для каждого изображения фото или рисунок?</p>
+      <form class="game__content">
         <div class="game__option">
-          <img src="http://placehold.it/705x455" alt="Option 1" width="705" height="455">
-          <label class="game__answer  game__answer--photo">
+          <img src="http://placehold.it/468x458" alt="Option 1" width="468" height="458">
+          <label class="game__answer game__answer--photo">
             <input class="visually-hidden" name="question1" type="radio" value="photo">
             <span>Фото</span>
           </label>
-          <label class="game__answer  game__answer--paint">
+          <label class="game__answer game__answer--paint">
             <input class="visually-hidden" name="question1" type="radio" value="paint">
+            <span>Рисунок</span>
+          </label>
+        </div>
+        <div class="game__option">
+          <img src="http://placehold.it/468x458" alt="Option 2" width="468" height="458">
+          <label class="game__answer  game__answer--photo">
+            <input class="visually-hidden" name="question2" type="radio" value="photo">
+            <span>Фото</span>
+          </label>
+          <label class="game__answer  game__answer--paint">
+            <input class="visually-hidden" name="question2" type="radio" value="paint">
             <span>Рисунок</span>
           </label>
         </div>
@@ -41,23 +52,20 @@ export default function insertGuessForOne() {
         <li class="stats__result stats__result--slow"></li>
         <li class="stats__result stats__result--fast"></li>
         <li class="stats__result stats__result--correct"></li>
-        <li class="stats__result stats__result--wrong"></li>
         <li class="stats__result stats__result--unknown"></li>
-        <li class="stats__result stats__result--slow"></li>
         <li class="stats__result stats__result--unknown"></li>
-        <li class="stats__result stats__result--fast"></li>
+        <li class="stats__result stats__result--unknown"></li>
+        <li class="stats__result stats__result--unknown"></li>
+        <li class="stats__result stats__result--unknown"></li>
         <li class="stats__result stats__result--unknown"></li>
       </ul>
     </section>`);
 
   const gameForm = node.querySelector(`.game__content`);
-  const answers = gameForm.querySelectorAll(`.game__answer`);
   gameForm.addEventListener(`change`, () => {
-    answers.forEach((item) => {
-      if (item.querySelector(`input:checked`)) {
-        makeScreenActive(insertFindImgOrPhoto());
-      }
-    });
+    if (gameForm.querySelector(`input[name="question1"]:checked`) && gameForm.querySelector(`input[name="question2"]:checked`)) {
+      makeScreenActive(insertGuessForOne());
+    }
   });
 
   const backButton = node.querySelector(`button.back`);
